+++ packages/yourloops/test/integration/assert/header.ts
@@ -29,8 +29,18 @@ import { type BoundFunctions, type queries, screen, within } from '@testing-libr
 import { UserRole } from '../../../lib/auth/models/enums/user-role.enum'
 import { type Team } from '../../../lib/team'
 import userEvent from '@testing-library/user-event'
+import { PRIVATE_TEAM_NAME } from '../../../lib/team/team.hook'
 
-const PRIVATE_TEAM_NAME = 'private'
+interface TeamMenuInfo {
+  selectedTeamName?: string
+  isSelectedTeamPrivate?: boolean
+  availableTeams: Team[]
+}
+
+export interface HeaderInfo {
+  loggedInUserFullName: string
+  teamMenuInfo: TeamMenuInfo
+}
 
 const checkHeader = (header: BoundFunctions<typeof queries>) => {
   expect(header.getByLabelText('YourLoops Logo')).toBeVisible()
@@ -65,9 +75,9 @@ const checkUserMenu = async (header: BoundFunctions<typeof queries>, userName: s
   expect(screen.queryByTestId('user-menu')).not.toBeInTheDocument()
 }
 
-const checkTeamScopeMenu = async (header: BoundFunctions<typeof queries>, selectedTeamParams: { teamName: string, isPrivate?: boolean }, availableTeams: Team[]) => {
-  const teamScopeMenuIcon = selectedTeamParams.isPrivate ? 'private-practice-icon' : 'medical-team-icon'
-  const teamScopeMenuText = selectedTeamParams.isPrivate ? 'My private practice' : selectedTeamParams.teamName
+const checkTeamScopeMenu = async (header: BoundFunctions<typeof queries>, teamMenuInfo: TeamMenuInfo) => {
+  const teamScopeMenuIcon = teamMenuInfo.isSelectedTeamPrivate ? 'private-practice-icon' : 'medical-team-icon'
+  const teamScopeMenuText = teamMenuInfo.isSelectedTeamPrivate ? 'My private practice' : teamMenuInfo.selectedTeamName
 
   expect(header.getByLabelText('Open team selection menu')).toBeVisible()
   expect(header.getByTestId(teamScopeMenuIcon)).toBeVisible()
@@ -78,7 +88,7 @@ const checkTeamScopeMenu = async (header: BoundFunctions<typeof queries>, select
   const teamScopeMenu = within(screen.getByTestId('team-scope-menu'))
   expect(teamScopeMenu.getByText('My private practice')).toBeVisible()
   expect(teamScopeMenu.getByText('Care teams')).toBeVisible()
-  availableTeams.forEach((team: Team) => {
+  teamMenuInfo.availableTeams.forEach((team: Team) => {
     if (team.name === PRIVATE_TEAM_NAME) {
       return
     }
@@ -90,18 +100,18 @@ const checkTeamScopeMenu = async (header: BoundFunctions<typeof queries>, select
   expect(screen.queryByTestId('team-scope-menu')).not.toBeInTheDocument()
 }
 
-export const checkHcpHeader = async (fullName: string, selectedTeamParams: { teamName: string, isPrivate?: boolean }, availableTeams: Team[]) => {
-  const header = within(screen.getByTestId('app-main-header'))
+export const checkHcpHeader = async (headerInfo: HeaderInfo) => {
+  const header = within(await screen.findByTestId('app-main-header'))
 
   expect(header.getByText('Patients')).toBeVisible()
-  if (selectedTeamParams.isPrivate) {
+  if (headerInfo.teamMenuInfo.isSelectedTeamPrivate) {
     expect(header.queryByText('Care team')).not.toBeInTheDocument()
   } else {
     expect(header.getByText('Care team')).toBeVisible()
   }
 
-  await checkTeamScopeMenu(header, selectedTeamParams, availableTeams)
-  await checkUserMenu(header, fullName, UserRole.Hcp)
+  await checkTeamScopeMenu(header, headerInfo.teamMenuInfo)
+  await checkUserMenu(header, headerInfo.loggedInUserFullName, UserRole.Hcp)
   checkHeader(header)
 }
 
@@ -119,4 +129,12 @@ export const checkPatientHeader = async (fullName: string) => {
 
   await checkUserMenu(header, fullName, UserRole.Patient)
   checkHeader(header)
+}
+
+export const changeTeamScope = async (currentTeamName: string, wantedTeamName: string) => {
+  const header = within(screen.getByTestId('app-main-header'))
+  await userEvent.click(header.getByText(currentTeamName))
+
+  const teamScopeMenu = within(screen.getByTestId('team-scope-menu'))
+  await userEvent.click(teamScopeMenu.getByText(wantedTeamName))
 }