+++ packages/yourloops/test/integration/assert/layout.ts
@@ -25,22 +25,47 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { checkCaregiverHeader, checkHcpHeader, checkPatientHeader } from './header'
-import { checkFooter } from './footer'
-import { UserRole } from '../../../lib/auth/models/enums/user-role.enum'
-import { type Team } from '../../../lib/team'
+import { checkCaregiverHeader, checkPatientHeader } from './header'
+import { screen, within } from '@testing-library/react'
+import { checkFooterForCaregiver, checkFooterForPatient } from './footer'
 
-export const checkHCPLayout = async (fullName: string, selectedTeamParams: { teamName: string, isPrivate?: boolean }, availableTeams: Team[], needFooterLanguageSelector: boolean = false) => {
-  await checkHcpHeader(fullName, selectedTeamParams, availableTeams)
-  checkFooter({ role: UserRole.Hcp, needFooterLanguageSelector })
+export interface PatientDashboardLayoutParams {
+  isChatCardVisible: boolean
+  isMedicalFilesCardVisible: boolean
+  isMonitoringAlertCardVisible: boolean
 }
 
 export const checkCaregiverLayout = async (fullName: string, needFooterLanguageSelector?: true) => {
   await checkCaregiverHeader(fullName)
-  checkFooter({ role: UserRole.Caregiver, needFooterLanguageSelector })
+  checkFooterForCaregiver(needFooterLanguageSelector)
 }
 
 export const checkPatientLayout = async (fullName: string, needFooterLanguageSelector?: true) => {
   await checkPatientHeader(fullName)
-  checkFooter({ role: UserRole.Patient, needFooterLanguageSelector })
+  checkFooterForPatient(needFooterLanguageSelector)
+}
+
+export const checkPatientDashboardLayout = async (patientDashboardLayout: PatientDashboardLayoutParams) => {
+  const dashboard = within(await screen.findByTestId('patient-dashboard'))
+  expect(dashboard.getByText('Data calculated on the last 7 days')).toBeVisible()
+  expect(dashboard.getByText('Patient statistics')).toBeVisible()
+  expect(dashboard.getByText('Device Usage')).toBeVisible()
+
+  if (patientDashboardLayout.isChatCardVisible) {
+    expect(dashboard.queryByTestId('chat-card')).toBeVisible()
+  } else {
+    expect(dashboard.queryByTestId('chat-card')).not.toBeInTheDocument()
+  }
+
+  if (patientDashboardLayout.isMedicalFilesCardVisible) {
+    expect(dashboard.queryByTestId('medical-files-card')).toBeVisible()
+  } else {
+    expect(dashboard.queryByTestId('medical-files-card')).not.toBeInTheDocument()
+  }
+
+  if (patientDashboardLayout.isMonitoringAlertCardVisible) {
+    expect(dashboard.queryByTestId('monitoring-alerts-card')).toBeVisible()
+  } else {
+    expect(dashboard.queryByTestId('monitoring-alerts-card')).not.toBeInTheDocument()
+  }
 }