+++ packages/yourloops/test/integration/assert/monitoring-alerts.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,18 +25,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { useTranslation } from 'react-i18next'
+import { screen } from '@testing-library/react'
 
-interface UserNameHookReturn {
-  getUserName: (firstName: string, lastName: string, fullName: string) => string
-}
-
-export const useUserName = (): UserNameHookReturn => {
-  const { t } = useTranslation('yourloops')
-
-  const getUserName = (firstName: string, lastName: string, fullName: string): string => {
-    return firstName && lastName ? t('user-name', { firstName, lastName }) : fullName
-  }
-
-  return { getUserName }
+export const checkMonitoringAlertsCard = async () => {
+  const monitoringAlertsCard = screen.getByTestId('monitoring-alerts-card')
+  expect(monitoringAlertsCard).toBeVisible()
+  expect(monitoringAlertsCard).toHaveTextContent('EventsCurrent eventsTime spent out of the target range10%Severe hypoglycemia20%Data not transferred30%')
 }