+++ packages/yourloops/test/integration/use-cases/monitoring-alerts-parameters-management.ts
@@ -25,26 +25,17 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { useUserName } from './user-name.hook'
-import { type GridComparatorFn } from '@mui/x-data-grid'
-import { type Patient } from '../patient/models/patient.model'
+import {
+  checkMonitoringAlertsDialogContentMgdl, checkMonitoringAlertsDialogContentMmol,
+  checkMonitoringAlertsDialogDefaultButtonMgdl, checkMonitoringAlertsDialogSaveButtonMmol
+} from '../assert/monitoring-alerts-parameters'
 
-interface SortComparatorsHookReturn {
-  sortByUserName: SortByUserNameComparator
+export const testMonitoringAlertsParametersConfigurationDialogMgdl = async (): Promise<void> => {
+  await checkMonitoringAlertsDialogContentMgdl()
+  await checkMonitoringAlertsDialogDefaultButtonMgdl()
 }
 
-interface SortByUserNameComparator extends GridComparatorFn<Patient> {
-  (patient1: Patient, patient2: Patient): number
-}
-
-export const useSortComparatorsHook = (): SortComparatorsHookReturn => {
-  const { getUserName } = useUserName()
-
-  const sortByUserName: SortByUserNameComparator = (patient1, patient2): number => {
-    const patient1FullName = getUserName(patient1.profile.firstName, patient1.profile.lastName, patient1.profile.fullName)
-    const patient2FullName = getUserName(patient2.profile.firstName, patient2.profile.lastName, patient2.profile.fullName)
-    return patient1FullName.localeCompare(patient2FullName)
-  }
-
-  return { sortByUserName }
+export const testMonitoringAlertsParametersConfigurationDialogMmol = async (): Promise<void> => {
+  await checkMonitoringAlertsDialogContentMmol()
+  await checkMonitoringAlertsDialogSaveButtonMmol()
 }