+++ packages/yourloops/test/integration/use-cases/patients-management.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,20 +25,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-const commonJestConfig = require('../common-jest.config')
-module.exports = {
-  ...commonJestConfig,
+import {
+  checkAddPatientMedicalTeamDialogContent,
+  checkAddPatientMedicalTeamDialogInvite,
+  checkAddPatientPrivateButtonTooltip,
+  checkRemovePatientMedicalTeamConfirm,
+  checkRemovePatientMedicalTeamDialogContent,
+  checkRemovePatientMedicalTeamError,
+  checkRemovePatientPrivateConfirm,
+  checkRemovePatientPrivateDialogContent
+} from '../assert/patient-list'
 
-  bail: true,
+export const testPatientManagementPrivateTeam = async () => {
+  await checkAddPatientPrivateButtonTooltip()
+  await checkRemovePatientPrivateDialogContent()
+  await checkRemovePatientPrivateConfirm()
+}
 
-  displayName: 'yourloops integration',
-
-  maxWorkers: 4,
-
-  // The glob patterns Jest uses to detect test files
-  testMatch: [
-    '<rootDir>/**/*.spec.tsx'
-  ],
-
-  testTimeout: 150000
+export const testPatientManagementMedicalTeam = async () => {
+  await checkRemovePatientMedicalTeamDialogContent()
+  await checkRemovePatientMedicalTeamConfirm()
+  await checkRemovePatientMedicalTeamError()
+  await checkAddPatientMedicalTeamDialogContent()
+  await checkAddPatientMedicalTeamDialogInvite()
 }